+++ simulation/simulation_interface/src/conversions.cpp
@@ -12,6 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
+#include <scenario_simulator_exception/exception.hpp>
 #include <simulation_interface/conversions.hpp>
 #include <string>
 #include <vector>
@@ -258,7 +259,7 @@ void toProto(const openscenario_msgs::msg::EntityType & type, openscenario_msgs:
     return;
   }
   std::string message = "type of the Entity Type is inavlid!\ntype is " + std::to_string(type.type);
-  THROW_PROTOBUF_PARAMETER_ERROR(message);
+  THROW_SIMULATION_ERROR(message);
 }
 
 void toMsg(const openscenario_msgs::EntityType & proto, openscenario_msgs::msg::EntityType & type)
@@ -276,7 +277,7 @@ void toMsg(const openscenario_msgs::EntityType & proto, openscenario_msgs::msg::
     return;
   }
   std::string message = "type of the Entity Type is inavlid!";
-  THROW_PROTOBUF_PARAMETER_ERROR(message);
+  THROW_SIMULATION_ERROR(message);
 }
 
 void toProto(