+++ simulation/traffic_simulator/include/traffic_simulator/entity/entity_manager.hpp
@@ -30,11 +30,11 @@
 #include <openscenario_msgs/msg/entity_status_with_trajectory_array.hpp>
 #include <openscenario_msgs/msg/vehicle_parameters.hpp>
 #include <rclcpp/rclcpp.hpp>
+#include <scenario_simulator_exception/exception.hpp>
 #include <stdexcept>
 #include <string>
 #include <traffic_simulator/entity/ego_entity.hpp>
 #include <traffic_simulator/entity/entity_base.hpp>
-#include <traffic_simulator/entity/exception.hpp>
 #include <traffic_simulator/entity/pedestrian_entity.hpp>
 #include <traffic_simulator/entity/vehicle_entity.hpp>
 #include <traffic_simulator/hdmap_utils/hdmap_utils.hpp>
@@ -191,7 +191,7 @@ public:
   try {                                                                        \
     return entities_.at(name)->IDENTIFIER(std::forward<decltype(xs)>(xs)...);  \
   } catch (const std::out_of_range &) {                                        \
-    throw SimulationRuntimeError("Unknown entity " + name + " was specified"); \
+    THROW_SEMANTIC_ERROR("entity : ", name, "does not exist");                 \
   }                                                                            \
   static_assert(true, "")
 
@@ -315,7 +315,7 @@ public:
       result.first->second->setTrafficLightManager(traffic_light_manager_ptr_);
       return result.second;
     } else {
-      throw traffic_simulator::SimulationRuntimeError("Entity '" + name + "' is already exists.");
+      THROW_SEMANTIC_ERROR("entity : ", name, " is already exists.");
     }
   }