+++ simulation/traffic_simulator/include/traffic_simulator/entity/vehicle_entity.hpp
@@ -23,7 +23,6 @@
 #include <traffic_simulator/behavior/vehicle/behavior_tree.hpp>
 #include <traffic_simulator/behavior/vehicle/lane_change_action.hpp>
 #include <traffic_simulator/entity/entity_base.hpp>
-#include <traffic_simulator/entity/exception.hpp>
 #include <traffic_simulator/entity/vehicle_parameter.hpp>
 
 // headers in pugixml